pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
]
